+++ jvagent/action/cockpit/routing/__init__.py
@@ -1,1 +1,0 @@
-"""Pre-cockpit posture classification and routing types."""